+++ test/nebulex/adapters/multilevel_inclusive_test.exs
@@ -11,6 +11,7 @@ defmodule Nebulex.Adapters.MultilevelInclusiveTest do
     ]
 
   use Nebulex.DistributedTest
+  use Nebulex.Distributed.TransactionTest
   use Nebulex.MultilevelTest
   use Nebulex.MultilevelQueryableTest
   use Nebulex.MultilevelInfoTest